apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FlowForge Backend - AI agents that design, build, and run workflows from natural language"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.3",
    "uvicorn>=0.40.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "claude-agent-sdk>=0.1.31",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
